fix: checkbox의 check 표시가 화면에 보여지지 않는 문제 해결
checkbox를 checkbox button으로 변경하여 문제 해결
