+++ src/components/common/index.ts
@@ -1,6 +1,6 @@
 export { default as BackButton } from './BackButton/BackButton';
 export { default as BottomButton } from './BottomButton/BottomButton';
-export { default as Checkbox } from './Checkbox/Checkbox';
+export { default as CheckBoxButton } from './CheckboxButton/CheckBoxButton';
 export { default as CountButton } from './CountButton/CountButton';
 export { default as Navigation } from './Navigation/Navigation';
 export { default as DeleteButton } from './DeleteButton/DeleteButton';
