+++ src/components/shoppingCart/CartList/CartList.tsx
@@ -1,5 +1,5 @@
 import { CartItem } from '@appTypes/shoppingCart';
-import { Checkbox } from '@components/common';
+import { CheckBoxButton } from '@components/common';
 import { CartListItem } from '@components/shoppingCart';
 import { useCheckCartItem } from '@hooks/shoppingCart';
 import { cartItemsSelector } from '@recoil/shoppingCart';
@@ -18,7 +18,7 @@ const CartList: React.FC<CartListProps> = () => {
   return (
     <div>
       <Styled.CartListButtonGroup>
-        <Checkbox checked={isAllChecked} onChange={(event) => onCheckAllCartItems(event.target.checked)} />
+        <CheckBoxButton checked={isAllChecked} onClick={() => onCheckAllCartItems(!isAllChecked)} />
         <span className="label">전체 선택</span>
       </Styled.CartListButtonGroup>
       <Styled.CartItemContainer>
